requirements.txt
playwright==1.41.0
pytest==7.4.4
openai==1.10.0
python-dotenv==1.0.1
sqlalchemy==2.0.25
pydantic==2.6.0
aiohttp==3.9.3
beautifulsoup4==4.12.3
fake-useragent==1.4.0
email-validator==2.1.0.post1
pandas==2.2.0
openpyxl==3.1.2
tenacity==8.2.3
